Moved AI classes to sperate folder (ollama4j)
